algorithms/requirements.txt
numpy>=1.24
Pillow>=10.0
scipy>=1.11
reedsolo>=1.7
pytest>=7.4
